audio_to_text/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Django==4.2.4
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
langdetect==1.0.9
libretranslatepy==2.1.1
lxml==4.9.3
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
translate==3.6.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
